Remove SQLEditorMenu component - this was a new component started at the beginning to refactor stuff but was dropped off, removed to prevent confusion
